requirement.txt
absl-py==0.8.1
AnyQt==0.0.10
appdirs==1.4.4
asgiref==3.2.3
astor==0.8.0
atomicwrites==1.4.0
attrs==19.3.0
backcall==0.1.0
backports.functools-lru-cache==1.6.1
baycomp==1.0.2
beautifulsoup4==4.9.1
biopython==1.76
bleach==3.1.0
boto==2.49.0
boto3==1.13.11
botocore==1.16.11
Bottleneck==1.3.2
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.12.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
commonmark==0.9.1
contextvars==2.4
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
defusedxml==0.6.0
dictdiffer==0.8.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.3
django-allauth==0.41.0
django-cors-headers==3.2.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
dlib==19.8.1
docutils==0.15.2
docx2txt==0.8
easydict==1.9
entrypoints==0.3
et-xmlfile==1.0.1
face-recognition==1.3.0
face-recognition-models==0.3.0
future==0.18.2
gast==0.3.2
gensim==3.8.3
grpcio==1.19.0
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
h5py==2.9.0
hpack==3.0.0
hstspreload==2020.5.5
html5lib==0.9999999
httpx==0.12.1
hyperframe==5.2.0
idna==2.8
imageio==2.8.0
immutables==0.12
importlib-metadata==1.6.0
imutil==0.2.5
imutils==0.5.3
ipykernel==5.1.3
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jdcal==1.4.1
jedi==0.13.3
Jinja2==2.10.3
jmespath==0.10.0
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
Keras==2.2.5
Keras-Applications==1.0.8
keras-datasets==0.1.0
Keras-Preprocessing==1.1.0
keyring==21.2.1
keyrings.alt==3.4.0
kiwisolver==1.0.1
livelossplot==0.4.1
lockfile==0.12.2
LTTL==2.0.9
lxml==4.5.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
mkl-fft==1.0.10
mkl-random==1.0.2
mock==2.0.0
mod-wsgi==4.7.1+ap24vc14
more-itertools==8.3.0
msgpack==1.0.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.4
nltk==3.4.5
notebook==5.2.2
numpy==1.18.0
oauthlib==3.1.0
odfpy==1.4.1
olefile==0.46
opencv-python==4.1.2.30
openpyxl==3.0.3
openTSNE==0.4.3
orange-canvas-core==0.1.12
orange-widget-base==4.5.0
Orange3==3.25.0
Orange3-Text==0.9.1
Orange3-Textable==3.1.5
Orange3-Timeseries==0.3.6
packaging==20.3
pandas==0.25.1
pandas-datareader==0.8.1
pandocfilters==1.4.2
parso==0.4.0
patsy==0.5.1
pbr==5.4.4
pdfminer3k==1.3.1
pickleshare==0.7.5
Pillow==6.1.0
pluggy==0.13.1
ply==3.11
prometheus-client==0.6.0
prompt-toolkit==1.0.15
protobuf==3.7.1
py==1.8.1
pycountry==19.8.18
pycparser==2.20
Pygments==2.5.2
PyJWT==1.7.1
pyparsing==2.4.0
pyqtgraph==0.10.0
pyreadline==2.1
pyrsistent==0.15.2
PySocks==1.7.1
pytest==5.4.2
python-datauri==0.2.8
python-dateutil==2.8.0
python-louvain==0.14
python3-openid==3.1.0
pytz==2019.2
PyWavelets==1.1.1
pywin32==223
pywin32-ctypes==0.2.0
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.6.0
requests==2.21.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
s3transfer==0.3.3
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.5.0
serverfiles==0.3.0
setuptools-git==1.2
simhash==1.9.0
simplegeneric==0.8.1
six==1.13.0
sklearn==0.0
smart-open==2.0.0
sniffio==1.1.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
soupsieve==2.0.1
sqlparse==0.3.0
static3==0.7.0
statsmodels==0.11.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.13.1
tensorflow-tensorboard==0.1.8
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.4
torch==1.2.0
torchvision==0.4.0
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.3
treetaggerwrapper==2.3
tweepy==3.8.0
ufal.udpipe==1.2.0.3
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wikipedia==1.4.0
wincertstore==0.2
wrapt==1.11.2
xgboost==0.90
xlrd==1.2.0
XlsxWriter==1.2.8
zipp==3.1.0
psycopg2==2.6.1